requirements.txt
ultralytics==8.0.21
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
requests>=2.31.0
tqdm>=4.65.0